// src/types/generated/contract.ts
export interface InputConfig {
  "focus-follows-mouse"?: boolean | null;
  keyboard?: KeyboardConfig | null;
  "mod-key"?: string | null;
  mouse?: MouseConfig | null;
  tablet?: TabletConfig | null;
  touch?: TouchConfig | null;
  touchpad?: TouchpadConfig | null;
  trackpoint?: TrackpointConfig | null;
  "warp-mouse-to-focus"?: boolean | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "KeyboardConfig".
 */
export interface KeyboardConfig {
  numlock?: boolean | null;
  "repeat-delay"?: number | null;
  "repeat-rate"?: number | null;
  "track-layout"?: string | null;
  xkb?: XkbConfig | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "XkbConfig".
 */
export interface XkbConfig {
  file?: string | null;
  layout?: string | null;
  model?: string | null;
  options?: string | null;
  rules?: string | null;
  variant?: string | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "MouseConfig".
 */
export interface MouseConfig {
  "accel-profile"?: string | null;
  "accel-speed"?: number | null;
  "left-handed"?: boolean | null;
  "natural-scroll"?: boolean | null;
  off?: boolean | null;
  "scroll-button"?: string | null;
  "scroll-method"?: string | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "TabletConfig".
 */
export interface TabletConfig {
  off?: boolean | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "TouchConfig".
 */
export interface TouchConfig {
  off?: boolean | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "TouchpadConfig".
 */
export interface TouchpadConfig {
  "accel-profile"?: string | null;
  "accel-speed"?: number | null;
  drag?: boolean | null;
  dwt?: boolean | null;
  dwtp?: boolean | null;
  "left-handed"?: boolean | null;
  "middle-emulation"?: boolean | null;
  "natural-scroll"?: boolean | null;
  off?: boolean | null;
  "scroll-button"?: string | null;
  "scroll-method"?: string | null;
  tap?: boolean | null;
}
/**
 * This interface was referenced by `Contract`'s JSON-Schema
 * via the `definition` "TrackpointConfig".
 */
export interface TrackpointConfig {
  "accel-profile"?: string | null;
  "accel-speed"?: number | null;
  "left-handed"?: boolean | null;
  "natural-scroll"?: boolean | null;
  off?: boolean | null;
  "scroll-button"?: string | null;
  "scroll-method"?: string | null;
}